+++ bin/tests.sh
@@ -27,10 +27,8 @@ verbose="v -v -v"
 
 declare -a testing
 # testing=("schedule")
-# testing=("discount" "expiration" "message" "poll" "post" "schedule")
-testing=("discount" "message" "upload")
-testing=("message")
-testing=("upload")
+# testing=("message" "upload")
+testing=("discount" "expiration" "message" "poll" "post" "schedule")
 
 echo "##################################################" >> ../onlysnarf/logs/tests.txt 2>&1
 echo "################## Start #########################" >> ../onlysnarf/logs/tests.txt 2>&1